Weather-app/requirements.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
billiard==4.2.0
cattrs==23.2.3
celery==5.4.0
certifi==2024.7.4
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.0.7
django-bootstrap5==22.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
exceptiongroup==1.2.2
fake-useragent==1.5.1
flatbuffers==24.3.25
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.7.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
kombu==5.3.7
numpy==2.0.0
openmeteo_requests==1.2.0
openmeteo_sdk==1.11.11
pandas==2.2.2
platformdirs==4.2.2
prompt_toolkit==3.0.47
PyJWT==2.8.0
python-dateutil==2.9.0.post0
pytz==2024.1
redis==5.0.7
requests==2.32.3
requests-cache==1.2.1
retry-requests==2.0.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
url-normalize==1.4.3
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13